Update README.md
The word "needs" should be replaced with "requires" to make the sentence more grammatically correct. It should be:

"video_reader - This requires ffmpeg to be installed and torchvision to be built from source. There shouldn't be any conflicting version of ffmpeg installed. Currently, this is only supported on Linux."

This change improves the clarity and accuracy of the sentence.